Python/microRobotFramework_python_02/requirements.txt
pyserial>=3.5
typing>=3.7.4
argparse>=1.1
